aggregated groups and channels
