+++ src/test/elements/ciscospark/groups.js
@@ -4,6 +4,7 @@ const suite = require('core/suite');
 const cloud = require('core/cloud');
 const payload = require('./assets/groups');
 const channelPayload = require('./assets/channels');
+const usersPayload = require('./assets/channelUsers');
 
 suite.forElement('collaboration', 'groups', { payload: payload }, (test) => {
   const opts = {
@@ -21,6 +22,21 @@ suite.forElement('collaboration', 'groups', { payload: payload }, (test) => {
       .then(r => groupId = r.body.id)
       .then(r => cloud.post(`${test.api}/${groupId}/channels`, channelPayload))
       .then(r => cloud.get(`${test.api}/${groupId}/channels`))
+      .then(r => cloud.delete(`${test.api}/${groupId}`));
+  });
+  it(`should allow CRUDS for ${test.api}/:id/users`, () => {
+    let groupId, id;
+    let userUpdatedPayload = {
+      isModerator: false
+    };
+    return cloud.post(test.api, payload)
+      .then(r => groupId = r.body.id)
+      .then(r => cloud.post(`${test.api}/${groupId}/users`, usersPayload))
+      .then(r => id = r.body.id)
+      .then(r => cloud.get(`${test.api}/${groupId}/users/${id}`))
+      .then(r => cloud.get(`${test.api}/${groupId}/users`))
+      .then(r => cloud.put(`${test.api}/${groupId}/users/${id}`, userUpdatedPayload))
+      .then(r => cloud.delete(`${test.api}/${groupId}/users/${id}`))
       .then(r => cloud.delete(`${test.api}/${groupId}`));
   });
 });
